Funciona creacion usuario, funciona creacion de deporte, funciona creacion de partido, funciona busqueda de partido por proximidad

+++ tp/src/main/java/com/g4/tp/repository/IMatchRepository.java
@@ -4,9 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.g4.tp.model.entities.Match;
 
-public interface IMatchRepository extends JpaRepository<Match, Long> {
-    // TODO: Agregar métodos de búsqueda específicos como:
-    // List<Match> findBySport(String sport);
-    // List<Match> findByState(String state);
+public interface IMatchRepository extends JpaRepository<Match, Integer> {
+
 
 }
