+++ tp/src/main/java/com/g4/tp/repository/ISportRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.g4.tp.model.entities.Sport;
 
-public interface ISportRepository extends JpaRepository<Sport, Long> {
+public interface ISportRepository extends JpaRepository<Sport, Integer> {
     Sport findByName(String name);
 
 }
