+++ tp/src/main/java/com/g4/tp/repository/IUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.g4.tp.model.entities.User;
 
-public interface IUserRepository extends JpaRepository<User, Long> {
+public interface IUserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
